refactor: wip reworking store
