+++ src/reducers/rootReducer.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux";
+import websiteReducer from "./websiteReducer";
+
+const rootReducer = combineReducers({
+  websiteReducer,
+});
+
+export default rootReducer;
